Revert "Move the Forger class into its own module"
